renamed gnomad_parser_genomes.py to gnomad_v2_parser.py; deleted gnomad_parser_exomes.py since the generic functions in gnomad_v2_parser.py can read both genome and exome data; changed the imports accordingly in the uploader
